sahur_poc/sahur/sahur-core/requirements.txt
pydantic>=2.0.0
langchain
langgraph
websockets
openai>=1.0.0
anthropic>=0.5.0
python-dotenv>=1.0.0
